remove Points Reopened Before and Points Closed Before
History calcs for these values are wrong and unnecessary
If a Task is Reopened, its points will be the same as when it
was closed and any changes will still be picked up in the points
change transaction.

If a Task is closed before the Sprint start, its points should not
be added to the points closed in the past and brought forward to the
present, because points closed is only used to determine current
points remaining.  There is no value in knowing how many points
were remaining in a Sprint that did not exist yet.

Change-Id: I4ada671839518b8a60bd6dc1a861868045488c01