overrides /tag/ routes for all projects
this is a bold commit that overrides
the project board controllers in favor of Sprint

the behavior and appearance of the boards vary depending on whether
the project has been enabled as a Sprint or not

As far as is possible, the Sprint controllers are identical to
Projects.  The main difference is the conditional logic that
allows different views.

The primary benefit is navigation clarity for Sprint users.
Also, this removes the Sprint UI Event Listener which is not
needed since the routing now occurs at the board controller

includes updated ProjectProfileController

Change-Id: I57fd399a495d991801632640824f6413bf231750