Fixes board issues caused by SprintValidator
Updates SprintBoardTaskEditController
Phabricator does not pass the current project context
to the task edit controller.  Since validation occurs at the card level,
rendering the card after a form post requires a new project object.
The project object does not need to be the current project, but it has to be
a Sprint in order for the board to be re-rendered

The assumption is that all cards in a Sprint Board will have at least
one valid Sprint project attachment, so this should never fail.

Also fixes missing default and update routes in SprintApplication

Change-Id: Ibcc25d3b0dac5e9b8f4a5809c2d8d12b226cd6f6